+++ hc/api/tests/test_notify_whatsapp.py
@@ -7,6 +7,7 @@ from datetime import timedelta as td
 from unittest.mock import patch
 
 from django.core import mail
+from django.test.utils import override_settings
 from django.utils.timezone import now
 
 from hc.api.models import Channel, Check, Notification
@@ -27,6 +28,7 @@ class NotifyWhatsAppTestCase(BaseTestCase):
         self.channel.save()
         self.channel.checks.add(self.check)
 
+    @override_settings(TWILIO_FROM="+000", TWILIO_MESSAGING_SERVICE_SID=None)
     @patch("hc.api.transports.curl.request")
     def test_it_works(self, mock_post):
         mock_post.return_value.status_code = 200
@@ -34,8 +36,8 @@ class NotifyWhatsAppTestCase(BaseTestCase):
 
         self.channel.notify(self.check)
 
-        args, kwargs = mock_post.call_args
-        payload = kwargs["data"]
+        payload = mock_post.call_args.kwargs["data"]
+        self.assertEqual(payload["From"], "whatsapp:+000")
         self.assertEqual(payload["To"], "whatsapp:+1234567890")
 
         n = Notification.objects.get()
@@ -45,6 +47,18 @@ class NotifyWhatsAppTestCase(BaseTestCase):
         # sent SMS counter should go up
         self.profile.refresh_from_db()
         self.assertEqual(self.profile.sms_sent, 1)
+
+    @override_settings(TWILIO_MESSAGING_SERVICE_SID="dummy-sid")
+    @patch("hc.api.transports.curl.request")
+    def test_it_uses_messaging_service(self, mock_post):
+        mock_post.return_value.status_code = 200
+        self._setup_data()
+
+        self.channel.notify(self.check)
+
+        payload = mock_post.call_args.kwargs["data"]
+        self.assertEqual(payload["MessagingServiceSid"], "dummy-sid")
+        self.assertFalse("From" in payload)
 
     @patch("hc.api.transports.curl.request")
     def test_it_obeys_up_down_flags(self, mock_post):
@@ -87,6 +101,5 @@ class NotifyWhatsAppTestCase(BaseTestCase):
 
         self.channel.notify(self.check)
 
-        args, kwargs = mock_post.call_args
-        payload = kwargs["data"]
+        payload = mock_post.call_args.kwargs["data"]
         self.assertIn("Foo > Bar & Co", payload["Body"])