+++ static/js/checks.js
@@ -69,25 +69,28 @@ $(function () {
     function applyFilters() {
         // Make a list of currently checked tags:
         var checked = [];
-        var qs = [];
+        var url = new URL(window.location.href);
+        url.search = "";
         $("#my-checks-tags .checked").each(function(index, el) {
             checked.push(el.textContent);
-            qs.push({"name": "tag", "value": el.textContent});
+            url.searchParams.append("tag", el.textContent);
         });
 
         var search = $("#search").val().toLowerCase();
         if (search) {
-            qs.push({"name": "search", "value": search});
+            url.searchParams.append("search", search);
         }
 
         // Update hash
         if (window.history && window.history.replaceState) {
-            var url = $("#checks-table").data("list-url");
-            if (qs.length) {
-                url += "?" + $.param(qs);
-            }
-            window.history.replaceState({}, "", url);
+            window.history.replaceState({}, "", url.toString());
         }
+
+        // Update sort links
+        document.querySelectorAll("a[data-sort-value]").forEach((a) => {
+            url.searchParams.set("sort", a.dataset.sortValue);
+            a.setAttribute("href", url.toString());
+        });
 
         // No checked tags and no search string: show all
         if (checked.length == 0 && !search) {
@@ -139,7 +142,6 @@ $(function () {
         window.location = url;
         return false;
     });
-
 
     $(".pause").tooltip({
         title: "Pause this check?<br />Click again to confirm.",